Generic method type arguments
Lookahead until the end of the current production trying to match a type. (pattern-matching is simple but works for test cases so far)
If it looks like a type, we can parse the type.
Otherwise we fall back to normal expression parsing and assume < is less-than comparison operator.
System isn't perfect, but should give us what we need for simple usage.
Improvements saved for future work.