Update to ParserObjects v0.2.2
GrammarStringifyVisitor now lives as BnfStringifyVisitor in ParserObjects.
Remove Find and Replace visitors, that functionality is moved to ParserObjects
Update SeparatedList() usages to match what it is in ParserObjects

One test is failing, we need to consider a way to have a fallback production in RightApplyParser if we match left and middle but not right.